+++ openm/libopenm/msg/msgExecBase.cpp
@@ -43,7 +43,7 @@ int MsgExecBase::groupRank(void) const
 * @param[in] i_size      size of array
 * @param[in] i_valueArr  value array to send
 */
-void MsgExecBase::startSend(int i_sendTo, MsgTag i_msgTag, const type_info & i_type, long long i_size, void * i_valueArr)
+void MsgExecBase::startSend(int i_sendTo, MsgTag i_msgTag, const type_info & i_type, size_t i_size, void * i_valueArr)
 {
     try {
         lock_guard<recursive_mutex> lck(msgMutex);
@@ -97,7 +97,7 @@ void MsgExecBase::startSendPacked(int i_sendTo, const IRowBaseVec & i_rowVec, co
 * @param[in]     i_size      size of array
 * @param[in,out] io_valueArr allocated buffer to recieve value array
 */
-void MsgExecBase::startRecv(int i_recvFrom, MsgTag i_msgTag, const type_info & i_type, long long i_size, void * io_valueArr)
+void MsgExecBase::startRecv(int i_recvFrom, MsgTag i_msgTag, const type_info & i_type, size_t i_size, void * io_valueArr)
 {
     try {
         lock_guard<recursive_mutex> lck(msgMutex);
@@ -151,7 +151,7 @@ void MsgExecBase::startRecvPacked(int i_recvFrom, IRowBaseVec & io_resultRowVec,
 * @param[in]     i_size      size of array
 * @param[in,out] io_valueArr allocated buffer to recieve value array
 */
-bool MsgExecBase::tryReceive(int i_recvFrom, MsgTag i_msgTag, const type_info & i_type, long long i_size, void * io_valueArr) const
+bool MsgExecBase::tryReceive(int i_recvFrom, MsgTag i_msgTag, const type_info & i_type, size_t i_size, void * io_valueArr) const
 {
     try {
         lock_guard<recursive_mutex> lck(msgMutex);
