+++ openm/libopenm/msg/msgMpiSend.cpp
@@ -74,7 +74,7 @@ void MpiSendBase::cleanup(void) throw()
 * @param[in] i_valueArr value array to send
  */
 MpiSendArray::MpiSendArray(
-    int i_selfRank, int i_sendTo, MsgTag i_msgTag, const type_info & i_type, long long i_size, void * i_valueArr
+    int i_selfRank, int i_sendTo, MsgTag i_msgTag, const type_info & i_type, size_t i_size, void * i_valueArr
     ) :
     MpiSendBase(i_selfRank, i_sendTo, i_msgTag)
 {
